fix: delete account

+++ app/Http/Controllers/UserController.php
@@ -47,7 +47,8 @@ class UserController extends Controller
     }
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
-        return to_route('web.data_master.index');   
+        return to_route('web.data_master.index');
     }
 }
